+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Changelog for the gruf-profiler gem.
 
+h3. 0.1.0
+
+- Add rubocop
+- Add bundler-audit
+- Add configuration options for logging for memory profiler
+
 h3. 0.0.1
 
 - Initial public release